+++ specula/data_objects/gaussian_convolution_kernel.py
@@ -55,7 +55,7 @@ class GaussianConvolutionKernel(ConvolutionKernel):
     def calculate_lgs_map(self):
         self.real_kernels = lgs_map_sh(
             self.dimx, self.pupil_size_m, 0, 90e3, [0], profz=[1.0], fwhmb=self.spot_size, ps=self.pxscale,
-            ssp=self.dimension, overs=1, theta=self.lgs_tt, xp=self.xp)
+            ssp=self.dimension, overs=1, theta=self.lgs_tt, dtype=self.dtype, xp=self.xp)
 
         self.process_kernels(return_fft=self.return_fft)
